Add first layout
It introduces the new bashrc layout. Instead of load each file
(variables, functions, aliases and so on) it load all files of each
directory contained into $CONFIGURATION/conf.d. Those directories are
loaded in crescent order, thus the lower layer must have a lower number.
For instance, 1-var.d contains files that exports environment variables
that could be used for the underlying layers.

Each file must not depend on a structure that are contained in the same
layer, otherwise it could create a cyclic dependency.

The new layout stands for:
- 1-vars.d: environment variables;
- 2-lib.d: bash functions which are an API;
- 3-functions.d: bash functions that should be used by the user;
- 4-aliases.d: similar to functions but they contains files with aliases
that does not have arguments;
- 5-exec.d: statements that are executed when bashrc is loaded;

Signed-off-by: r1cm3d <ricardo.medeiros@pismo.io>