Fix 'Last Command Bug'
It fixes last command bug. This was a known bug that used to occurs when
one press Ctrl+p or arrow key up in terminal listing the last commands.
At the certain moment, the text of one of the commands stays in the
screen making difficult to understand where is the cursor.

This bug was caused by a wrong color set on `$GIT_COLOR` environment
variable.

Signed-off-by: r1cm3d <ricardo.medeiros@pismo.io>